Implement non-buffered streaming for step logs
The implementation is a bit of a hack until AnnotatedLargeText in
Jenkins-core has an interface-based constructor, or some other method
for passing in a dynamic log file.

Now that the log-index is read incrementally while streaming, the length
of the step log needs to be determined before returning. The only option
for determining the step logs length is by reading the full index file.
A new IndexReader class abstracts the parsing of the index. It will be
used by both the full read in advance and the incremental read while
streaming the log content.